pyproject.toml
[project]
name = "movie-rating"
version = "0.1.0"
description = ""
authors = [
    {name = "Arman Hosseini",email = "armanhosseini878787@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "fastapi (>=0.127.0,<0.128.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "movie_rating", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
